Change license to CC-BY-SA
The current license is not clear and a LICENSE file is missing. I
recommend to use CC-BY-SA to be able to include other material licensed
unter this license, especially Wikimedia projects. CC-BY and CC0 may
also be good choices but they do not allow merging in CC-BY-SA content.
Anyway a license with NC clause as indicated by the current icons should
be avoided (see <https://freedomdefined.org/Licenses/NC> for arguments)!